--- composeApp/src/commonMain/kotlin/org/tavioribeiro/calculatordpi/core/extension/String.kt
@@ -1,3 +1,0 @@
-package org.tavioribeiro.calculatordpi.core.extension
-
-
