--- composeApp/src/commonMain/kotlin/org/tavioribeiro/calculatordpi/general_utils/DirectoryPicker.kt
@@ -1,10 +1,0 @@
-package org.tavioribeiro.calculatordpi.general_utils
-
-import androidx.compose.runtime.Composable
-
-@Composable
-expect fun DirectoryPicker(
-    show: Boolean,
-    title: String,
-    onResult: (String?) -> Unit
-)
