--- composeApp/src/commonMain/kotlin/org/tavioribeiro/calculatordpi/presentation/components/inputs/InputEnums.kt
@@ -1,8 +1,0 @@
-package org.tavioribeiro.calculatordpi.presentation.components.inputs
-
-enum class InputTypeEnum {
-    TEXT,
-    NUMBER,
-    TEXT_PASSWORD,
-    NUMBER_PASSWORD
-}
