--- composeApp/src/commonMain/kotlin/org/tavioribeiro/calculatordpi/presentation/features/main/tabs/AiAgentsTab.kt
@@ -1,56 +1,0 @@
-package org.tavioribeiro.calculatordpi.presentation.features.main.tabs
-
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.safeContentPadding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import org.tavioribeiro.calculatordpi.presentation.components.buttons.ActionButton
-import org.tavioribeiro.calculatordpi.presentation.components.inputs.MidInput
-import org.tavioribeiro.calculatordpi.theme.AppTheme
-import org.tavioribeiro.calculatordpi.theme.ThemeState
-import androidx.compose.runtime.getValue
-import androidx.compose.ui.layout.ContentScale
-import calculatordpi.composeapp.generated.resources.Res
-import calculatordpi.composeapp.generated.resources.icon_clear
-import calculatordpi.composeapp.generated.resources.icon_folder
-import calculatordpi.composeapp.generated.resources.icon_history
-import calculatordpi.composeapp.generated.resources.icon_plus
-import calculatordpi.composeapp.generated.resources.icon_robot
-import org.jetbrains.compose.resources.painterResource
-import org.tavioribeiro.calculatordpi.presentation.components.buttons.IconTextButton
-import org.tavioribeiro.calculatordpi.presentation.components.buttons.UnderlineButton
-
-
-@Composable
-fun AiAgentsTab() {
-    Column(
-        Modifier.height(391.dp)
-            .fillMaxWidth()
-            .background(AppTheme.colors.color2)
-            .padding(horizontal = 25.dp),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-
-    }
-}
