--- composeApp/src/commonMain/kotlin/org/tavioribeiro/calculatordpi/theme/ThemeState.kt
@@ -1,14 +1,0 @@
-package org.tavioribeiro.calculatordpi.theme
-
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.setValue
-
-object ThemeState {
-    var currentColors by mutableStateOf(darkColors)
-        private set
-
-    fun toggleTheme() {
-        currentColors = if (currentColors == darkColors) lightColors else darkColors
-    }
-}
