+++ composeApp/src/commonMain/kotlin/org/tavioribeiro/commitic/general_utils/DirectoryPicker.kt
@@ -1,0 +1,10 @@
+package org.tavioribeiro.commitc.general_utils
+
+import androidx.compose.runtime.Composable
+
+@Composable
+expect fun DirectoryPicker(
+    show: Boolean,
+    title: String,
+    onResult: (String?) -> Unit
+)
