+++ composeApp/src/commonMain/kotlin/org/tavioribeiro/commitic/presentation/components/buttons/ActionButton.kt
@@ -1,0 +1,45 @@
+package org.tavioribeiro.commitc.presentation.components.buttons
+
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import org.tavioribeiro.commitc.theme.AppTheme
+
+@Composable
+fun ActionButton(
+    text: String,
+    onClick: () -> Unit,
+){
+    Button(
+        onClick = onClick,
+        colors = ButtonDefaults.buttonColors(
+            containerColor = AppTheme.colors.color7,
+            contentColor = AppTheme.colors.onColor7
+        ),
+        contentPadding = PaddingValues(
+            start = 20.dp,
+            top = 12.dp,
+            end = 20.dp,
+            bottom = 12.dp
+        ),
+        shape = RoundedCornerShape(4.dp)
+    ) {
+        Text(
+            text = text,
+            modifier = Modifier.padding(0.dp),
+            color = AppTheme.colors.onColor7,
+            style = MaterialTheme.typography.labelLarge.copy(
+                fontWeight = FontWeight.Bold
+            )
+        )
+    }
+}
