+++ composeApp/src/commonMain/kotlin/org/tavioribeiro/commitic/presentation/components/buttons/UnderlineButton.kt
@@ -1,0 +1,102 @@
+package org.tavioribeiro.commitc.presentation.components.buttons
+
+import androidx.compose.animation.animateColorAsState
+import androidx.compose.animation.core.tween
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.hoverable
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.interaction.collectIsHoveredAsState
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.ripple.rememberRipple
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.painter.Painter
+import androidx.compose.ui.semantics.Role
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import org.tavioribeiro.commitc.theme.AppTheme
+
+@Composable
+fun UnderlineButton(
+    text: String,
+    onClick: () -> Unit,
+    icon: Painter,
+    isSelected: Boolean,
+    modifier: Modifier = Modifier
+) {
+    val interactionSource = remember { MutableInteractionSource() }
+    val isHovered by interactionSource.collectIsHoveredAsState()
+
+    val targetColor = when {
+        isSelected -> AppTheme.colors.color7
+        isHovered -> AppTheme.colors.color6
+        else -> AppTheme.colors.color5
+    }
+
+    val contentColor by animateColorAsState(
+        targetValue = targetColor,
+        animationSpec = tween(durationMillis = 200),
+        label = "contentColorAnimation"
+    )
+
+    Column(
+        modifier = modifier
+            .clip(MaterialTheme.shapes.small)
+            .hoverable(interactionSource = interactionSource)
+            .clickable(
+                onClick = onClick,
+                role = Role.Button,
+                indication = null,
+                interactionSource = interactionSource
+            )
+            .padding(top = 8.dp, bottom = 2.dp, start = 16.dp, end = 16.dp),
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        Row(
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Icon(
+                painter = icon,
+                contentDescription = null,
+                tint = contentColor,
+                modifier = Modifier.padding(end = 8.dp)
+            )
+            Text(
+                text = text,
+                color = contentColor,
+                style = MaterialTheme.typography.titleMedium.copy(
+                    fontWeight = FontWeight.SemiBold
+                )
+            )
+        }
+
+        Spacer(modifier = Modifier.height(6.dp))
+
+        Box(
+            modifier = Modifier
+                .width(130.dp)
+                .height(2.dp)
+                .background(if (isSelected) contentColor else Color.Transparent)
+        )
+    }
+}
