+++ composeApp/src/commonMain/kotlin/org/tavioribeiro/commitic/presentation/components/inputs/InputEnums.kt
@@ -1,0 +1,8 @@
+package org.tavioribeiro.commitc.presentation.components.inputs
+
+enum class InputTypeEnum {
+    TEXT,
+    NUMBER,
+    TEXT_PASSWORD,
+    NUMBER_PASSWORD
+}
