+++ composeApp/src/commonMain/kotlin/org/tavioribeiro/commitic/theme/AppTheme.kt
@@ -1,0 +1,30 @@
+package org.tavioribeiro.commitc.theme
+
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.ReadOnlyComposable
+import androidx.compose.runtime.staticCompositionLocalOf
+
+val localAppColors = staticCompositionLocalOf { darkColors }
+
+object AppTheme {
+    val colors: AppColors
+        @Composable
+        @ReadOnlyComposable
+        get() = localAppColors.current
+}
+
+@Composable
+fun MyAppTheme(
+    content: @Composable () -> Unit
+) {
+    val currentColors = ThemeState.currentColors
+
+    CompositionLocalProvider(localAppColors provides currentColors) {
+        MaterialTheme(
+            content = content,
+            typography = FigtreeTypography()
+        )
+    }
+}
