--- composeApp/src/jvmMain/kotlin/org/tavioribeiro/calculatordpi/general_utils/DirectoryPicker.kt
@@ -1,39 +1,0 @@
-package org.tavioribeiro.calculatordpi.general_utils
-
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
-import javax.swing.JFileChooser
-import javax.swing.UIManager
-import javax.swing.filechooser.FileSystemView
-
-@Composable
-actual fun DirectoryPicker(
-    show: Boolean,
-    title: String,
-    onResult: (String?) -> Unit
-) {
-    LaunchedEffect(show) {
-        if (show) {
-            try {
-                UIManager.setLookAndFeel(UIManager.getSystemLookAndFeelClassName())
-            } catch (e: Exception) {
-
-            }
-
-            val fileChooser = JFileChooser(FileSystemView.getFileSystemView().homeDirectory).apply {
-                dialogTitle = title
-                fileSelectionMode = JFileChooser.DIRECTORIES_ONLY
-                isAcceptAllFileFilterUsed = false
-            }
-
-            val result = fileChooser.showOpenDialog(null)
-            if (result == JFileChooser.APPROVE_OPTION) {
-                onResult(fileChooser.selectedFile.absolutePath)
-            } else {
-                onResult(null)
-            }
-        }
-    }
-}
