+++ composeApp/src/jvmMain/kotlin/org/tavioribeiro/commitic/screen_utils/ScreenSizeInfo.kt
@@ -1,0 +1,17 @@
+package org.tavioribeiro.commitc.screen_utils
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.platform.LocalWindowInfo
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+
+
+@Composable
+actual fun getScreenWidthDp(): Dp {
+        return LocalWindowInfo.current.containerSize.width.dp
+}
+
+@Composable
+actual fun getScreenHeightDp(): Dp {
+    return LocalWindowInfo.current.containerSize.height.dp
+}
