--- composeApp/src/wasmJsMain/kotlin/org/tavioribeiro/calculatordpi/Platform.wasmJs.kt
@@ -1,7 +1,0 @@
-package org.tavioribeiro.calculatordpi
-
-class WasmPlatform: Platform {
-    override val name: String = "Web with Kotlin/Wasm"
-}
-
-actual fun getPlatform(): Platform = WasmPlatform()
