--- composeApp/src/wasmJsMain/kotlin/org/tavioribeiro/calculatordpi/general_utils/DirectoryPicker.kt
@@ -1,14 +1,0 @@
-package org.tavioribeiro.calculatordpi.general_utils
-
-import androidx.compose.runtime.Composable
-
-@Composable
-actual fun DirectoryPicker (
-    show: Boolean,
-    title: String,
-    onResult: (String?) -> Unit
-) {
-    if (show) {
-        onResult(null)
-    }
-}
