--- composeApp/src/wasmJsMain/kotlin/org/tavioribeiro/calculatordpi/main.kt
@@ -1,12 +1,0 @@
-package org.tavioribeiro.calculatordpi
-
-import androidx.compose.ui.ExperimentalComposeUiApi
-import androidx.compose.ui.window.ComposeViewport
-import kotlinx.browser.document
-
-@OptIn(ExperimentalComposeUiApi::class)
-fun main() {
-    ComposeViewport(document.body!!) {
-        App()
-    }
-}
