+++ composeApp/src/wasmJsMain/kotlin/org/tavioribeiro/commitic/Platform.wasmJs.kt
@@ -1,0 +1,7 @@
+package org.tavioribeiro.commitc
+
+class WasmPlatform: Platform {
+    override val name: String = "Web with Kotlin/Wasm"
+}
+
+actual fun getPlatform(): Platform = WasmPlatform()
