+++ composeApp/src/wasmJsMain/kotlin/org/tavioribeiro/commitic/main.kt
@@ -1,0 +1,12 @@
+package org.tavioribeiro.commitc
+
+import androidx.compose.ui.ExperimentalComposeUiApi
+import androidx.compose.ui.window.ComposeViewport
+import kotlinx.browser.document
+
+@OptIn(ExperimentalComposeUiApi::class)
+fun main() {
+    ComposeViewport(document.body!!) {
+        App()
+    }
+}
